add ability to pull commit history from github
Using the "Import History" button, commits can be taken directly from
the web interface, which can then be modified. As a result of this
commit, the canvas has been resized to be 53 cells wide, and the cell
size modified to be 12 to better reflect the way it will appear on
github.com.

Signed-off-by: Ricky Ayoub <vgmoose@gmail.com>