more accurate grid script
The script that is generated now accurately represents what the graph
will look like by accounting for the day light savings time shift and
the correct number of weeks. It also uses a new algorithm to determine
how many commits to make each day according to this formula:

0 :  0 commits
1 :  (1/4)*max - 1 commits
2 :  (2/4)*max - 1 commits
3 :  (3/4)*max - 1 commits
4 :  max commits

Where max by default is 10, and the numbers to the left of the colon
represent the five shades of commit colors.

Signed-off-by: Ricky Ayoub <vgmoose@gmail.com>